update OfJSON to provide compatibility with PHP versions earlier than 5.3.
Yeah, I'm an idiot and forgot to check to see if json_last_error() was available earlier.  Oh well, we're still using exception error handling if that's the case.
Should now work in PHP 5.2 and such at least, since it will just ignore the switch, and proceed directly to the else statement.