Changed ROOT_PATH to OF_ROOT, a more specific and useful constant for the Framework.
In order for any file in the framework to be loaded, we must define OF_ROOT as the path from __FILE__ to the framework root (src/ directory). This should allow us to autoload stuff within the framework itself and store Twig and Doctrine in vendor while maintaining a valid path.