Overhaul the core, provide config handling (and namespacing), dependency injection, and object management
Note that the config handling is not to be used in the application itself.  It is primarily intended to be used with dependency injection; any actual configurations should be handled through other means, such as with the Config utility (formerly OfConfig).
Again, application settings should not be stored via the Framework core.  Do that yourself some other way.